// interfaces/order.ts
import { ISize } from "./products";
import { IUser } from "./user";

export interface IOrder {

  _id?: string;
  user?: IUser | string;
  orderItems: IOrderItem[];
  shippingAddres :ShippingAddres;
  paymentResult?: string;

  numberOfOrderItems: number;
  subTotal: number;
  tax: number;
  total: number;

  isPaid: boolean;
  paidAt?: string;
  transactionId?: string;

}


export interface IOrderItem {
  _id: string;
  title: string;
  size: ISize;
  quantity: number;
  slug: string;
  image: string;
  price: number;
  gender: string;
}


export interface ShippingAddres {
  firstName: string;
  lastName: string;
  address: string;
  address2? : string;
  zip: string;
  city: string;
  country: string;
  phone: string;
} 